requirements.txt
python-a2s==1.4.0
aiohttp
flask
async-timeout
